+++ MANIFEST
@@ -1,12 +1,13 @@
 # file GENERATED by distutils, do NOT edit
-README.txt
 setup.py
-controls\Controls.json
-controls\GenerateControlTable.py
-controls\YawAngles.py
-controls\__init__.py
-utilities\AlfanoChebyshev.py
-utilities\AlfanoLib.py
-utilities\AlfanoStateEqLib.py
-utilities\__init__.py
-utilities\propellant_plots.py
+alfano\__init__.py
+alfano\controls\Controls.json
+alfano\controls\GenerateControlTable.py
+alfano\controls\YawAngleProxy.py
+alfano\controls\YawAngles.py
+alfano\controls\__init__.py
+alfano\utilities\AlfanoChebyshev.py
+alfano\utilities\AlfanoLib.py
+alfano\utilities\AlfanoStateEqLib.py
+alfano\utilities\__init__.py
+alfano\utilities\propellant_plots.py
