+++ MANIFEST.in
@@ -1,0 +1,13 @@
+# file GENERATED by distutils, do NOT edit
+setup.py
+EULA.pdf
+alfano\__init__.py
+alfano\controls\Controls.json
+alfano\controls\GenerateControlTable.py
+alfano\controls\YawAngles.py
+alfano\controls\__init__.py
+alfano\utilities\AlfanoChebyshev.py
+alfano\utilities\AlfanoLib.py
+alfano\utilities\AlfanoStateEqLib.py
+alfano\utilities\__init__.py
+alfano\utilities\propellant_plots.py
