+++ setup.py
@@ -2,17 +2,20 @@
 """
 Created on Sun Apr 28 15:08:37 2019
 
-@author: Colin
+@author: Colin Helms
 """
 
 from distutils.core import setup
 setup(name='alfano',
-      version='0.2a',
+      version='0.3a',
+      long_description='Libraries for trajectory solutions to Edelbaum-Alfano low-thrust orbit transfer.',
+      description='Edelbaum-Alfano libraries.',
       author='Colin Helms',
       author_email='colinhelms@outlook.com',
-      url='https://www.FreelanceRocketScience.com/downloads',
-      packages=['utilities', 'controls'],
-      package_data={'controls' : ['Controls.json']},
+      url='https://www.FreelanceRocketScience.com',
+      packages=['', 'alfano', 'alfano.utilities', 'alfano.controls'],
+      package_data={'alfano.controls' : ['Controls.json']},
+      license='EULA',
       classifiers=[
               'Development Status :: 2 - Alpha',
               'Environment :: Console',
